Rewrite message matching function in Lua scripts
As explained in #74, 'matching' is more suitable description of the
operations performed by this function than 'compare'. It also inverts
the boolean output of the function, which, given its new name, is only
natural.

Also, fixes a bug in cfg/basic.lua, which went unnoticed due to
circumstantial simplifications of that script: arguments order was
reversed, but it didn't matter until the pattern (second parameter)
contains -1.

Fixes #74